Move Bluetooth & BluetoothState to scanning package
