+++ app/src/main/java/com/aconno/acnsensa/BluetoothScanningService.kt
@@ -15,7 +15,7 @@ import com.aconno.acnsensa.dagger.bluetoothscanning.DaggerBluetoothScanningServi
 import com.aconno.acnsensa.data.http.EmptyPublisher
 import com.aconno.acnsensa.data.http.RESTPublisher
 import com.aconno.acnsensa.data.mqtt.GoogleCloudPublisher
-import com.aconno.acnsensa.domain.Bluetooth
+import com.aconno.acnsensa.domain.scanning.Bluetooth
 import com.aconno.acnsensa.domain.Publisher
 import com.aconno.acnsensa.domain.ifttt.GooglePublish
 import com.aconno.acnsensa.domain.ifttt.RESTPublish
