+++ app/src/main/java/com/aconno/acnsensa/viewmodel/factory/BluetoothViewModelFactory.kt
@@ -3,7 +3,7 @@ package com.aconno.acnsensa.viewmodel.factory
 import android.app.Application
 import android.arch.lifecycle.ViewModel
 import com.aconno.acnsensa.BluetoothStateReceiver
-import com.aconno.acnsensa.domain.Bluetooth
+import com.aconno.acnsensa.domain.scanning.Bluetooth
 import com.aconno.acnsensa.viewmodel.BluetoothViewModel
 
 class BluetoothViewModelFactory(
